Revert "Feature#3/mtg support (#23)"
This reverts commit 644e253f2e3c5ffcc389bae1981db391b2f05606.
